requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
whitenoise==6.11.0
